feat: get genres from remote
