+++ data/src/main/java/com/madrid/data/dataSource/local/mappers/CategoryMappers.kt
@@ -1,12 +1,12 @@
 package com.madrid.data.dataSource.local.mappers
 
 import com.madrid.data.dataSource.local.entity.CategoryEntity
-import com.madrid.data.dataSource.remote.response.movie.MovieGenre
+import com.madrid.data.dataSource.remote.response.genre.MovieGenre
 
 
 fun MovieGenre.toCategoryEntity(): CategoryEntity {
     return CategoryEntity(
-        id = this.id ?: 0,
+        categoryId = this.id ?: 0,
         categoryTitle = this.name ?: "Unknown",
         searchCount = 0
     )
