+++ data/src/main/java/com/madrid/data/dataSource/remote/MovieApi.kt
@@ -3,6 +3,7 @@ package com.madrid.data.dataSource.remote
 import com.madrid.data.dataSource.remote.response.artist.ArtistDetailsResponse
 import com.madrid.data.dataSource.remote.response.artist.SearchArtistResponse
 import com.madrid.data.dataSource.remote.response.common.TrailerResponse
+import com.madrid.data.dataSource.remote.response.genre.GenresResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieCreditsResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieDetailsResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieReviewResponse
@@ -60,6 +61,8 @@ interface MovieApi {
         @Path("movie_id") movieId: Int
     ): SimilarMoviesResponse
 
+    @GET("genre/movie/list")
+    suspend fun getMovieGenres(): GenresResponse
 
     // Series
     @GET("search/tv")
@@ -99,6 +102,8 @@ interface MovieApi {
         @Query("page") page: Int
     ): SearchSeriesResponse
 
+    @GET("genre/tv/list")
+    suspend fun getSeriesGenres(): GenresResponse
 
     // Artist
     @GET("search/person")
