+++ data/src/main/java/com/madrid/data/repositories/remote/RemoteDataSource.kt
@@ -3,6 +3,7 @@ package com.madrid.data.repositories.remote
 import com.madrid.data.dataSource.remote.response.artist.ArtistDetailsResponse
 import com.madrid.data.dataSource.remote.response.artist.SearchArtistResponse
 import com.madrid.data.dataSource.remote.response.common.TrailerResponse
+import com.madrid.data.dataSource.remote.response.genre.GenresResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieCreditsResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieDetailsResponse
 import com.madrid.data.dataSource.remote.response.movie.MovieReviewResponse
@@ -83,5 +84,8 @@ interface RemoteDataSource {
     suspend fun getArtistById(artistId: Int): ArtistDetailsResponse
 
 
+
+    // region genres
+    suspend fun getMovieGenres(): GenresResponse
 
 }
